api/requirements.txt
fastapi==0.105.0
uvicorn==0.34.0
pandas==2.3.3
scikit-learn==1.8.0
numpy==2.4.2
pyarrow==22.0.0
pyaml>=25.1.0
